feature/minimal-integration
- refactor for tests
